feat: create Order

+++ src/infrastructure/orm/migration/1719325567987-create_table_order_product.ts
@@ -4,16 +4,17 @@ export class CreateTableOrderProduct1719325567987 implements MigrationInterface 
 
     public async up(queryRunner: QueryRunner): Promise<void> {
         await queryRunner.query(`
-          CREATE TABLE public.order_products (
-            order_products_id serial PRIMARY KEY,
-            order_id int NOT NULL,
-            product_ref int NOT NULL,
-            quantity int NOT NULL,
-            price int NOT NULL,
-            created_at timestamp without time zone DEFAULT now() NOT NULL,
-            updated_at timestamp without time zone DEFAULT now() NOT NULL,
-            FOREIGN KEY (order_id) REFERENCES public.order(order_id)
-          );
+        CREATE TABLE public.order_products (
+          order_products_id serial PRIMARY KEY,
+          order_id int NOT NULL,
+          product_ref int NOT NULL,
+          quantity int NOT NULL,
+          price numeric NOT NULL,
+          created_at timestamp without time zone DEFAULT now() NOT NULL,
+          updated_at timestamp without time zone DEFAULT now() NOT NULL,
+          FOREIGN KEY (order_id) REFERENCES public.order(order_id)
+        );
+        
     
           CREATE SEQUENCE public.order_products_id_seq
             AS integer
